backend/Api/configs/requirements.txt
aerich==0.3.3
aiomysql==0.0.20
aiosqlite==0.16.0
attrs==20.3.0
beautifulsoup4==4.9.3
boto3==1.16.21
botocore==1.19.21
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
ecdsa==0.14.1
fastapi==0.61.2
h11==0.11.0
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
iso8601==0.1.13
jmespath==0.10.0
jwt==1.1.0
packaging==20.4
passlib==1.7.4
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.2
PyMySQL==0.9.2
pyparsing==2.4.7
PyPika==0.44.0
pytest==6.1.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-jose==3.2.0
pytz==2020.4
requests==2.25.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
starlette==0.13.6
toml==0.10.2
tortoise-orm==0.16.18
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.12.2
xmltodict==0.12.0
zipp==3.4.0
